drivers: modem: ublox-sara-r4: select automatic registration before enabling radio
The existing behavior of the SARA-R4 driver issues AT+CFUN=1 before
AT+COPS=0,0. This prevents the SARA-422 (and possibly others from
registering until the next AT+CFUN=0 is issued after the timeout waiting
for the network. This may have something to do with the auto-
registration done at start up (which should be interruptable according
to Section 2.1.1 of UBX-17003787 - R26) and how rapidly the AT+COPS=0,0
is issued afterwards. It is unclear if this is related to the user
reselection (3GPP TS 23.122) behavior or another internal conflict on
the SARA.

The solution implemented is not DRY but is limited to runtime behavior
by KConfig emitting an empty string "" in autoconf.h for string KConfigs
which are unset.

Signed-off-by: Peter Maxwell Warasila <madmaxwell@soundcomesout.com>