drivers: modem: ublox-sara-r4: socket_dns_offload uses resolution pool
Previously the ublox-sara-r4 used a single instance of the required
structures for offload_getaddrinfo to return. This commit modifies this
behavior to use a thread safe pool of resolution structures instead. The
number of resolutions available is equal to the number of sockets to
ensure a resolution will be available for the lifetime of each
connection.

This code assumes that applications which previously used the sara-r4
driver call freeaddrinfo() to correctly free the resolution. If
applications did not correctly call freeaddrinfo() in their applciation
code, this driver will eventually run out of resolutions to issue.

This change was tested with a proprietary board using the U-Blox
SARA-R422M8S modem. Two simultaneous connections werere made. The first
held its resolution by not calling freeaddrinfo(). The other got a
resolution, made a connection, and then freed its resolution at regular
intervals around 100 times. This verifies that the resolutions are being
freed correctly and there are no starvation issues. Both connetions
worked consistently.

Signed-off-by: Peter Maxwell Warasila <madmaxwell@soundcomesout.com>